Refactor and create a separate MagicEvaluator
