+++ src/engine.rs
@@ -1,3 +1,4 @@
 pub mod datalog;
 pub(crate) mod index_storage;
 pub(crate) mod storage;
+pub(crate) mod magic_evaluator;
